+++ src/main/auth/services/auth-get-profile.service.ts
@@ -28,6 +28,7 @@ export class AuthGetProfileService {
       where,
       include: {
         notifications: true,
+        profilePicture: true,
       },
     });